iOS: map menu button to Options, start to Menu. Analog triggers
